// block-performance-1.0.0/src/library/x-jmeter-cloud-control/target-server-get/type.ts
export interface Input {
    target_server_id: string;
}

export interface OutputOk {
    target_server: {
        id: string;
        fake?: boolean;
        create_time: string;
        update_time: string;
        cloud_server: { aws_lightsail_instance_id: string };
        ssh_username: string;
        address_ipv4: string;
    };
}

export interface OutputFail {}

export interface Callback<R> {
    none: () => R;
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}